call ffmpeg function without gil
Calling FFmpeg funtion will trigger log collection. if the logs
are to be passed to the python buffer, it will trigger the acquisition
of the GIL. A deadlock my occur between the log collection lock
and the GIL. Therefore, use the nogil mode when calling the FFmpeg method
from python.